Changed build scripts to allow easy re-building versions with increasing build numbers

+++ build_greyhole.sh
@@ -42,8 +42,8 @@
 #######
 # Usage
 
-if [ $# != 1 ]; then
-	echo "Usage: $0 <version>"
+if [ $# -lt 1 -o $# -gt 2 ]; then
+	echo "Usage: $0 <version> [build number]"
 	exit 1
 fi
 
@@ -71,6 +71,12 @@ ANNOUNCE_EMAIL='releases-announce@greyhole.net'
 
 
 export VERSION=$1
+
+if [ $# -eq 2 ]; then
+	export BUILD_NUMBER="$2"
+else
+	export BUILD_NUMBER="1"
+fi
 
 # Clean unwanted files
 find . -name "._*" -delete
@@ -103,7 +109,9 @@ fi
 #########################################
 # Transfer files to HOST:PATH_TO_RELEASES
 
-scp release/greyhole*$VERSION.tar.gz ${HOST}:${PATH_TO_RELEASES}/.
+if [ "$BUILD_NUMBER" = "1" ]; then
+	scp release/greyhole*$VERSION.tar.gz ${HOST}:${PATH_TO_RELEASES}/.
+fi
 scp release/*greyhole-$VERSION-*.src.rpm ${HOST}:${PATH_TO_RELEASES}/rpm/src/.
 scp release/*greyhole-$VERSION-*.x86_64.rpm ${HOST}:${PATH_TO_RELEASES}/rpm/x86_64/.
 scp release/*greyhole-$VERSION-*.i386.rpm ${HOST}:${PATH_TO_RELEASES}/rpm/i386/.
@@ -115,9 +123,14 @@ scp release/greyhole-$VERSION-*.deb ${HOST}:${PATH_TO_RELEASES}/deb/.
 # Update YUM/APT repo data
 
 ssh ${HOST} ${PATH_TO_REPOS_UPDATER}/update_yum_repodata.sh
-ssh ${HOST} ${PATH_TO_REPOS_UPDATER}/update_deb_repodata.sh $VERSION
+
 # update_deb_repodata.sh needs to be called from an interactive session on $HOST, because it needs the GPG secret key passphrase to work!
-# A reminder to do that will be echoed at the end of this script.
+echo
+echo "*******************"
+echo "You now need to execute the following command in the SSH shell that will open:"
+echo "  ${PATH_TO_REPOS_UPDATER}/update_deb_repodata.sh $VERSION ; exit"
+echo "*******************"
+ssh ${HOST}
 
 
 ############################################################
@@ -148,57 +161,50 @@ chmod +x ~/greyhole/greyhole ~/greyhole/greyhole-dfree
 ####################
 # Tag the git branch
 git clone git@github.com:gboudreau/Greyhole.git /tmp/Greyhole.git
-(cd /tmp/Greyhole.git; git tag $VERSION; git push --tags)
+if [ "$BUILD_NUMBER" = "1" ]; then
+	(cd /tmp/Greyhole.git; git tag $VERSION; git push --tags)
+else
+	(cd /tmp/Greyhole.git; git tag $VERSION-$BUILD_NUMBER; git push --tags)
+fi
 rm -rf /tmp/Greyhole.git
 
 
-############################
-# Update and email CHANGELOG
+##################
+# Update CHANGELOG
 
-cd release
-	LAST_TGZ=`ls -1atr *.tar.gz | grep -v web-app | grep -v 'hda-' | grep -B 1 greyhole-$VERSION | head -1`
-	tar --wildcards -x "*/CHANGES" -f $LAST_TGZ
-	tar --wildcards -x "*/CHANGES" -f greyhole-$VERSION.tar.gz
-
-	diff -b */CHANGES | sed -e 's/^> /- /' | grep -v '^[0-9]*a[0-9]*\,[0-9]*$' > /tmp/gh_changelog
-
-	find . -type d -name "greyhole-*" -exec rm -rf {} \; > /dev/null 2>&1
-	find . -type d -name "hda-greyhole-*" -exec rm -rf {} \; > /dev/null 2>&1
+if [ "$BUILD_NUMBER" = "1" ]; then
+	cd release
+		LAST_TGZ=`ls -1atr *.tar.gz | grep -v web-app | grep -v 'hda-' | grep -B 1 greyhole-$VERSION | head -1`
+		tar --wildcards -x "*/CHANGES" -f $LAST_TGZ
+		tar --wildcards -x "*/CHANGES" -f greyhole-$VERSION.tar.gz
 	
-	# Update $CHANGELOG_URL
-	echo "What's new in $VERSION" > CHANGELOG
-	echo "--------------------" >> CHANGELOG
-	cat /tmp/gh_changelog >> CHANGELOG
-	echo >> CHANGELOG
-	curl -s "${CHANGELOG_URL}" >> CHANGELOG
-	scp CHANGELOG ${HOST}:${PATH_TO_RELEASES}/CHANGELOG
-cd ..
+		diff -b */CHANGES | sed -e 's/^> /- /' | grep -v '^[0-9]*a[0-9]*\,[0-9]*$' > /tmp/gh_changelog
+	
+		find . -type d -name "greyhole-*" -exec rm -rf {} \; > /dev/null 2>&1
+		find . -type d -name "hda-greyhole-*" -exec rm -rf {} \; > /dev/null 2>&1
+		
+		# Update $CHANGELOG_URL
+		echo "What's new in $VERSION" > CHANGELOG
+		echo "--------------------" >> CHANGELOG
+		cat /tmp/gh_changelog >> CHANGELOG
+		echo >> CHANGELOG
+		curl -s "${CHANGELOG_URL}" >> CHANGELOG
+		scp CHANGELOG ${HOST}:${PATH_TO_RELEASES}/CHANGELOG
+	cd ..
+fi
 
 
 ############################################
 # Send notifications to Twitter/FB/IRC/email
 
-/usr/local/bin/twitter set "New version available: $VERSION - ChangeLog: http://t.co/hZheYwg"
-/usr/local/bin/fbcmd PPOST Greyhole "New version available: $VERSION - Downloads: http://www.greyhole.net/download/ or just use your package manager to update." 'ChangeLog' "${CHANGELOG_URL}"
-./irc_notif.sh "New version available $VERSION - Changelog: http://www.greyhole.net/releases/CHANGELOG" $VERSION
-
-# Email
-cat > /tmp/gh_email <<EOF
-This is an automated email.
-
-New RPM & DEB packages were created from a new Greyhole build: greyhole-$VERSION.tar.gz
-You can find this new build at the usual http://www.greyhole.net/releases/
-  and in the APT and YUM repositories.
-
-Changes from the previous version are:
-EOF
-cat /tmp/gh_changelog >> /tmp/gh_email
-mail -s "New Greyhole build available: $VERSION" $ANNOUNCE_EMAIL < /tmp/gh_email
+if [ "$BUILD_NUMBER" = "1" ]; then
+	/usr/local/bin/twitter set "New version available: $VERSION - ChangeLog: http://t.co/hZheYwg"
+	/usr/local/bin/fbcmd PPOST Greyhole "New version available: $VERSION - Downloads: http://www.greyhole.net/download/ or just use your package manager to update." 'ChangeLog' "${CHANGELOG_URL}"
+	./irc_notif.sh "New version available $VERSION - Changelog: http://www.greyhole.net/releases/CHANGELOG" $VERSION
+else
+	/usr/local/bin/twitter set "New packages available: $VERSION-$BUILD_NUMBER. If you couldn't install the previous packages, try this one."
+	/usr/local/bin/fbcmd PPOST Greyhole "New packages available: $VERSION-$BUILD_NUMBER. If you couldn't install the previous packages, try this one." 'ChangeLog' "${CHANGELOG_URL}"
+	./irc_notif.sh "New packages available $VERSION-$BUILD_NUMBER. If you couldn't install the previous packages, try this one." $VERSION
+fi
 
 ###
-
-echo
-echo "*******************"
-echo "You now need to ssh ${HOST} and manually execute the following command:"
-echo "  ${PATH_TO_REPOS_UPDATER}/update_deb_repodata.sh $VERSION"
-echo "*******************"
